"Happy Numbers"

+++ Problem Solving/src/HappyNumber.java
@@ -1,5 +1,4 @@
 import java.util.HashSet;
-import java.util.Map;
 import java.util.Set;
 
 public class HappyNumber {
@@ -15,7 +14,7 @@ public class HappyNumber {
         while (true){
             int sum=0;
             while(n!=0){
-                sum += Math.pow(n%10, 2);
+                sum += (int) Math.pow(n%10, 2);
                 n/=10;
             }
             if (sum==1) return true;
